refactor: 불필요한 권한을 부여하는 접근제어자 수정 및 패키지 구조 변경
